Moved GenerateCodeButton into separate component, added two Storybook stories
